requirements.txt
fastapi
uvicorn
whisperx
moviepy
requests
openai
# torch==2.2.0
transformers==4.30.2
torch ==2.2.0